Filter API & Entitiy info api
Making more Artists API routes as well as changing a lot of GETS to POSTS as gets should not acess req body. Updated some views in SQL. Working on API for filter components for albums and artists. As well as ability so see songs / albums that artists "owns" and reviews and songs associated with album